Minor tidy up
