+++ lib/engine/round/base.rb
@@ -33,8 +33,8 @@ module Engine
         current_entity.player
       end
 
-      def can_act?(entity)
-        active_entities.include?(entity)
+      def active_entities
+        [@current_entity] + crowded_corps
       end
 
       def next_entity
@@ -50,8 +50,36 @@ module Engine
         action.entity.pass!
       end
 
+      def process_action(action)
+        entity = action.entity
+        return @log << action if action.is_a?(Action::Message)
+        raise GameError, 'Game has ended' if @game.finished
+
+        if action.is_a?(Action::EndGame)
+          @log << '-- Game ended by player --'
+          return @game.end_game!
+        end
+
+        raise GameError, "It is not #{entity.name}'s turn" unless can_act?(entity)
+
+        if action.pass?
+          log_pass(entity)
+          pass(action)
+          pass_processed(action)
+        else
+          _process_action(action)
+          action_processed(action)
+        end
+        change_entity(action)
+        action_finalized(action)
+      end
+
       def finished?
         @game.finished || @entities.all?(&:passed?)
+      end
+
+      def can_act?(entity)
+        active_entities.include?(entity)
       end
 
       def auction?
